Update 5-2.txt
The previous answer, I guess, was for the question before this one. But even with this new answer, there still seems to be problem, as the heading is not the same as given. This may be an issue on the side of Grok or on my side. Someone who is willing to check this out would be highly appreciated!